+++ src/components/Image/Image.style.tsx
@@ -1,17 +1,22 @@
 import styled from 'styled-components';
+import { ReturnImageProps } from './Image';
 
-const StyledImage = styled.img`
+const StyledImage = styled.img<ReturnImageProps>`
   display: inline-block;
-  width: 100%;
 
+  img {
+    transition-property: opacity;
+    transition-delay: 0.3s;
+  }
+
+  img[data-src],
   &.lazyload,
   &.lazyloading {
     opacity: 0;
-    transition: opacity 400ms;
   }
 
+  img,
   &.lazyloaded {
-    transition: opacity 400ms;
     opacity: 1;
   }
 `;