+++ src/components/Image/Image.tsx
@@ -1,59 +1,58 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
-import AspectRatio from '../AspectRatio/AspectRatio';
-
-import { buildSrcSet, buildImageObject } from './helper';
+import buildImageProps from './helper';
 
 import StyledImage from './Image.style';
 
 export type ImageProps = {
   alt?: string;
-  src: string;
-  /**
-   * srcset object, e.g. { 300vw: "path/to/image", 400vw: "path/to/image" }
-   */
-  srcset?: { [key: string]: string };
+  src?: string;
+  width?: number;
+  height?: number;
+  srcSet?: string;
   sizes?: string;
-  /**
-   * Defines if image should lazy load
-   */
+  role?: string;
   lazy?: boolean;
-  /**
-   * Add ratio (e.g. 16:9) to avoid unintended jumping while image loading
-   */
-  ratio?: string;
   className?: string;
+  media?: string;
+  as?: 'img' | 'source';
+  hasNativeLazyLoading?: boolean;
 };
 
-const Image: React.FC<ImageProps> = ({
-  alt,
-  src,
-  srcset,
-  sizes,
-  lazy = false,
-  ratio,
-  className
-}: ImageProps) => {
-  const { image, imageClass } = buildImageObject(
-    {
-      alt,
-      role: !alt ? 'presentation' : undefined,
-      sizes,
-      src,
-      srcSet: srcset ? buildSrcSet(srcset) : undefined
-    },
-    lazy
-  );
+export type ReturnImageProps = React.HTMLProps<HTMLImageElement>;
 
-  if (ratio) {
-    return (
-      <AspectRatio ratio={ratio} className={className || ''}>
-        <StyledImage className={imageClass} {...image} />
-      </AspectRatio>
-    );
+const Image: React.FC<ImageProps> = (props: ImageProps) => {
+  const [hasNativeLazyLoading, sethasNativeLazyLoading] = useState(false);
+  const { alt, as = 'img' } = props;
+
+  useEffect(() => {
+    if (
+      typeof window !== 'undefined' &&
+      typeof HTMLImageElement !== 'undefined' &&
+      'loading' in HTMLImageElement.prototype
+    ) {
+      sethasNativeLazyLoading(true);
+    }
+  }, []);
+
+  const imageProps: ImageProps = buildImageProps({
+    ...props,
+    role: !alt ? 'presentation' : undefined,
+    hasNativeLazyLoading
+  });
+
+  if (as === 'source') {
+    delete imageProps.alt;
+    delete imageProps.lazy;
+    delete imageProps.src;
+    delete imageProps.className;
+    delete imageProps.width;
+    delete imageProps.height;
+
+    return <source {...imageProps} />;
   }
 
-  return <StyledImage className={`${className} ${imageClass}`} {...image} />;
+  return <StyledImage {...imageProps} />;
 };
 
 export default Image;