+++ src/components/Picture/index.ts
@@ -1,2 +1,0 @@
-export { default, Source } from './Picture';
-export type { SourceProps, PictureProps } from './Picture';