+++ CHANGELOG.md
@@ -15,11 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added new Image Component
 - Changed components to typescript
 - Updated stories and jest snapshot tests for typescript
 
 ### Removed
 
+- Dedicated Picture component in favor of new Image Component
 - Removed some unused packages (e.g. prop-types)
 
 ## [0.1.0] - 2020-10-05